Acerto Signin

+++ db/migrate/20170311184151_add_fator_to_unidade.rb
@@ -1,5 +1,5 @@
 class AddFatorToUnidade < ActiveRecord::Migration[5.0]
   def change
-    add_column :unidades, :fator, :double
+    add_column :unidades, :fator, :float
   end
 end
